+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * v0.0.33
   * Fix "$@" combination for strings.
   * Fix formatting at readonly editor windows.
+  * "csharpfixformat.style.activateDefaultFormatterAfter" option deprecated (prepare to be default formatter).
 * v0.0.32
   * Fix foreach loop without parentheses.
 * v0.0.31